TP5 - CPainting.class
float--> int

+++ TP5 Etude et amelioration d'une application/painting-ants-master/src/main/java/org/polytechtours/javaperformance/tp/paintingants/CPainting.java
@@ -34,9 +34,11 @@ import java.awt.event.MouseListener;
 public class CPainting extends Canvas implements MouseListener {
   private static final long serialVersionUID = 1L;
   // matrice servant pour le produit de convolution
-  static private float[][] mMatriceConv9 = new float[3][3];
-  static private float[][] mMatriceConv25 = new float[5][5];
-  static private float[][] mMatriceConv49 = new float[7][7];
+//  initialisation de la matrice de convolution : 
+//  float --> int
+  static private int[][] mMatriceConv9 = new int[3][3];
+  static private int[][] mMatriceConv25 = new int[5][5];
+  static private int[][] mMatriceConv49 = new int[7][7];
   // Objet de type Graphics permettant de manipuler l'affichage du Canvas
   private Graphics mGraphics;
   // Objet ne servant que pour les bloc synchronized pour la manipulation du
@@ -139,46 +141,46 @@ public class CPainting extends Canvas implements MouseListener {
     /*
      * 1 2 1 2 4 2 1 2 1
      */
-    CPainting.mMatriceConv9[0][0] = 1 / 16f;
-    CPainting.mMatriceConv9[0][1] = 2 / 16f;
-    CPainting.mMatriceConv9[0][2] = 1 / 16f;
-    CPainting.mMatriceConv9[1][0] = 2 / 16f;
-    CPainting.mMatriceConv9[1][1] = 4 / 16f;
-    CPainting.mMatriceConv9[1][2] = 2 / 16f;
-    CPainting.mMatriceConv9[2][0] = 1 / 16f;
-    CPainting.mMatriceConv9[2][1] = 2 / 16f;
-    CPainting.mMatriceConv9[2][2] = 1 / 16f;
+    CPainting.mMatriceConv9[0][0] = 1 / 16;
+    CPainting.mMatriceConv9[0][1] = 2 / 16;
+    CPainting.mMatriceConv9[0][2] = 1 / 16;
+    CPainting.mMatriceConv9[1][0] = 2 / 16;
+    CPainting.mMatriceConv9[1][1] = 4 / 16;
+    CPainting.mMatriceConv9[1][2] = 2 / 16;
+    CPainting.mMatriceConv9[2][0] = 1 / 16;
+    CPainting.mMatriceConv9[2][1] = 2 / 16;
+    CPainting.mMatriceConv9[2][2] = 1 / 16;
 
     // initialisation de la matrice de convolution : lissage moyen sur 25
     // cases
     /*
      * 1 1 2 1 1 1 2 3 2 1 2 3 4 3 2 1 2 3 2 1 1 1 2 1 1
      */
-    CPainting.mMatriceConv25[0][0] = 1 / 44f;
-    CPainting.mMatriceConv25[0][1] = 1 / 44f;
-    CPainting.mMatriceConv25[0][2] = 2 / 44f;
-    CPainting.mMatriceConv25[0][3] = 1 / 44f;
-    CPainting.mMatriceConv25[0][4] = 1 / 44f;
-    CPainting.mMatriceConv25[1][0] = 1 / 44f;
-    CPainting.mMatriceConv25[1][1] = 2 / 44f;
-    CPainting.mMatriceConv25[1][2] = 3 / 44f;
-    CPainting.mMatriceConv25[1][3] = 2 / 44f;
-    CPainting.mMatriceConv25[1][4] = 1 / 44f;
-    CPainting.mMatriceConv25[2][0] = 2 / 44f;
-    CPainting.mMatriceConv25[2][1] = 3 / 44f;
-    CPainting.mMatriceConv25[2][2] = 4 / 44f;
-    CPainting.mMatriceConv25[2][3] = 3 / 44f;
-    CPainting.mMatriceConv25[2][4] = 2 / 44f;
-    CPainting.mMatriceConv25[3][0] = 1 / 44f;
-    CPainting.mMatriceConv25[3][1] = 2 / 44f;
-    CPainting.mMatriceConv25[3][2] = 3 / 44f;
-    CPainting.mMatriceConv25[3][3] = 2 / 44f;
-    CPainting.mMatriceConv25[3][4] = 1 / 44f;
-    CPainting.mMatriceConv25[4][0] = 1 / 44f;
-    CPainting.mMatriceConv25[4][1] = 1 / 44f;
-    CPainting.mMatriceConv25[4][2] = 2 / 44f;
-    CPainting.mMatriceConv25[4][3] = 1 / 44f;
-    CPainting.mMatriceConv25[4][4] = 1 / 44f;
+    CPainting.mMatriceConv25[0][0] = 1 / 44;
+    CPainting.mMatriceConv25[0][1] = 1 / 44;
+    CPainting.mMatriceConv25[0][2] = 2 / 44;
+    CPainting.mMatriceConv25[0][3] = 1 / 44;
+    CPainting.mMatriceConv25[0][4] = 1 / 44;
+    CPainting.mMatriceConv25[1][0] = 1 / 44;
+    CPainting.mMatriceConv25[1][1] = 2 / 44;
+    CPainting.mMatriceConv25[1][2] = 3 / 44;
+    CPainting.mMatriceConv25[1][3] = 2 / 44;
+    CPainting.mMatriceConv25[1][4] = 1 / 44;
+    CPainting.mMatriceConv25[2][0] = 2 / 44;
+    CPainting.mMatriceConv25[2][1] = 3 / 44;
+    CPainting.mMatriceConv25[2][2] = 4 / 44;
+    CPainting.mMatriceConv25[2][3] = 3 / 44;
+    CPainting.mMatriceConv25[2][4] = 2 / 44;
+    CPainting.mMatriceConv25[3][0] = 1 / 44;
+    CPainting.mMatriceConv25[3][1] = 2 / 44;
+    CPainting.mMatriceConv25[3][2] = 3 / 44;
+    CPainting.mMatriceConv25[3][3] = 2 / 44;
+    CPainting.mMatriceConv25[3][4] = 1 / 44;
+    CPainting.mMatriceConv25[4][0] = 1 / 44;
+    CPainting.mMatriceConv25[4][1] = 1 / 44;
+    CPainting.mMatriceConv25[4][2] = 2 / 44;
+    CPainting.mMatriceConv25[4][3] = 1 / 44;
+    CPainting.mMatriceConv25[4][4] = 1 / 44;
 
     // initialisation de la matrice de convolution : lissage moyen sur 49
     // cases
@@ -186,61 +188,61 @@ public class CPainting extends Canvas implements MouseListener {
      * 1 1 2 2 2 1 1 1 2 3 4 3 2 1 2 3 4 5 4 3 2 2 4 5 8 5 4 2 2 3 4 5 4 3 2 1 2
      * 3 4 3 2 1 1 1 2 2 2 1 1
      */
-    CPainting.mMatriceConv49[0][0] = 1 / 128f;
-    CPainting.mMatriceConv49[0][1] = 1 / 128f;
-    CPainting.mMatriceConv49[0][2] = 2 / 128f;
-    CPainting.mMatriceConv49[0][3] = 2 / 128f;
-    CPainting.mMatriceConv49[0][4] = 2 / 128f;
-    CPainting.mMatriceConv49[0][5] = 1 / 128f;
-    CPainting.mMatriceConv49[0][6] = 1 / 128f;
+    CPainting.mMatriceConv49[0][0] = 1 / 128;
+    CPainting.mMatriceConv49[0][1] = 1 / 128;
+    CPainting.mMatriceConv49[0][2] = 2 / 128;
+    CPainting.mMatriceConv49[0][3] = 2 / 128;
+    CPainting.mMatriceConv49[0][4] = 2 / 128;
+    CPainting.mMatriceConv49[0][5] = 1 / 128;
+    CPainting.mMatriceConv49[0][6] = 1 / 128;
 
-    CPainting.mMatriceConv49[1][0] = 1 / 128f;
-    CPainting.mMatriceConv49[1][1] = 2 / 128f;
-    CPainting.mMatriceConv49[1][2] = 3 / 128f;
-    CPainting.mMatriceConv49[1][3] = 4 / 128f;
-    CPainting.mMatriceConv49[1][4] = 3 / 128f;
-    CPainting.mMatriceConv49[1][5] = 2 / 128f;
-    CPainting.mMatriceConv49[1][6] = 1 / 128f;
+    CPainting.mMatriceConv49[1][0] = 1 / 128;
+    CPainting.mMatriceConv49[1][1] = 2 / 128;
+    CPainting.mMatriceConv49[1][2] = 3 / 128;
+    CPainting.mMatriceConv49[1][3] = 4 / 128;
+    CPainting.mMatriceConv49[1][4] = 3 / 128;
+    CPainting.mMatriceConv49[1][5] = 2 / 128;
+    CPainting.mMatriceConv49[1][6] = 1 / 128;
 
-    CPainting.mMatriceConv49[2][0] = 2 / 128f;
-    CPainting.mMatriceConv49[2][1] = 3 / 128f;
-    CPainting.mMatriceConv49[2][2] = 4 / 128f;
-    CPainting.mMatriceConv49[2][3] = 5 / 128f;
-    CPainting.mMatriceConv49[2][4] = 4 / 128f;
-    CPainting.mMatriceConv49[2][5] = 3 / 128f;
-    CPainting.mMatriceConv49[2][6] = 2 / 128f;
+    CPainting.mMatriceConv49[2][0] = 2 / 128;
+    CPainting.mMatriceConv49[2][1] = 3 / 128;
+    CPainting.mMatriceConv49[2][2] = 4 / 128;
+    CPainting.mMatriceConv49[2][3] = 5 / 128;
+    CPainting.mMatriceConv49[2][4] = 4 / 128;
+    CPainting.mMatriceConv49[2][5] = 3 / 128;
+    CPainting.mMatriceConv49[2][6] = 2 / 128;
 
-    CPainting.mMatriceConv49[3][0] = 2 / 128f;
-    CPainting.mMatriceConv49[3][1] = 4 / 128f;
-    CPainting.mMatriceConv49[3][2] = 5 / 128f;
-    CPainting.mMatriceConv49[3][3] = 8 / 128f;
-    CPainting.mMatriceConv49[3][4] = 5 / 128f;
-    CPainting.mMatriceConv49[3][5] = 4 / 128f;
-    CPainting.mMatriceConv49[3][6] = 2 / 128f;
+    CPainting.mMatriceConv49[3][0] = 2 / 128;
+    CPainting.mMatriceConv49[3][1] = 4 / 128;
+    CPainting.mMatriceConv49[3][2] = 5 / 128;
+    CPainting.mMatriceConv49[3][3] = 8 / 128;
+    CPainting.mMatriceConv49[3][4] = 5 / 128;
+    CPainting.mMatriceConv49[3][5] = 4 / 128;
+    CPainting.mMatriceConv49[3][6] = 2 / 128;
 
-    CPainting.mMatriceConv49[4][0] = 2 / 128f;
-    CPainting.mMatriceConv49[4][1] = 3 / 128f;
-    CPainting.mMatriceConv49[4][2] = 4 / 128f;
-    CPainting.mMatriceConv49[4][3] = 5 / 128f;
-    CPainting.mMatriceConv49[4][4] = 4 / 128f;
-    CPainting.mMatriceConv49[4][5] = 3 / 128f;
-    CPainting.mMatriceConv49[4][6] = 2 / 128f;
+    CPainting.mMatriceConv49[4][0] = 2 / 128;
+    CPainting.mMatriceConv49[4][1] = 3 / 128;
+    CPainting.mMatriceConv49[4][2] = 4 / 128;
+    CPainting.mMatriceConv49[4][3] = 5 / 128;
+    CPainting.mMatriceConv49[4][4] = 4 / 128;
+    CPainting.mMatriceConv49[4][5] = 3 / 128;
+    CPainting.mMatriceConv49[4][6] = 2 / 128;
 
-    CPainting.mMatriceConv49[5][0] = 1 / 128f;
-    CPainting.mMatriceConv49[5][1] = 2 / 128f;
-    CPainting.mMatriceConv49[5][2] = 3 / 128f;
-    CPainting.mMatriceConv49[5][3] = 4 / 128f;
-    CPainting.mMatriceConv49[5][4] = 3 / 128f;
-    CPainting.mMatriceConv49[5][5] = 2 / 128f;
-    CPainting.mMatriceConv49[5][6] = 1 / 128f;
+    CPainting.mMatriceConv49[5][0] = 1 / 128;
+    CPainting.mMatriceConv49[5][1] = 2 / 128;
+    CPainting.mMatriceConv49[5][2] = 3 / 128;
+    CPainting.mMatriceConv49[5][3] = 4 / 128;
+    CPainting.mMatriceConv49[5][4] = 3 / 128;
+    CPainting.mMatriceConv49[5][5] = 2 / 128;
+    CPainting.mMatriceConv49[5][6] = 1 / 128;
 
-    CPainting.mMatriceConv49[6][0] = 1 / 128f;
-    CPainting.mMatriceConv49[6][1] = 1 / 128f;
-    CPainting.mMatriceConv49[6][2] = 2 / 128f;
-    CPainting.mMatriceConv49[6][3] = 2 / 128f;
-    CPainting.mMatriceConv49[6][4] = 2 / 128f;
-    CPainting.mMatriceConv49[6][5] = 1 / 128f;
-    CPainting.mMatriceConv49[6][6] = 1 / 128f;
+    CPainting.mMatriceConv49[6][0] = 1 / 128;
+    CPainting.mMatriceConv49[6][1] = 1 / 128;
+    CPainting.mMatriceConv49[6][2] = 2 / 128;
+    CPainting.mMatriceConv49[6][3] = 2 / 128;
+    CPainting.mMatriceConv49[6][4] = 2 / 128;
+    CPainting.mMatriceConv49[6][5] = 1 / 128;
+    CPainting.mMatriceConv49[6][6] = 1 / 128;
 
     mSuspendu = false;
   }
